Corrected windows build issue For windows builds, the following error was produced:    XBHelpMenus.cpp(539): warning C4267: 'argument': conversion from 'size_t' to 'const _Ty', possible loss of data Which resulted later in an error.

+++ src/runtime_src/core/tools/common/XBHelpMenus.cpp
@@ -536,7 +536,7 @@ XBUtilities::create_suboption_list_string(const VectorPairStrings &_collection)
       pairs.first += "\'";     
     }
 
-    maxStringLength = std::max<unsigned int>(maxStringLength, pairs.first.length());
+    maxStringLength = std::max<unsigned int>(maxStringLength, static_cast<unsigned int>(pairs.first.length()));
   }
 
   const unsigned int indention = maxStringLength + 5;  // New line indention after the '-' character (5 extra spaces)
